migrate our portions of the file editor to webpack
