+++ apps/dashboard/app/assets/javascripts/application.js
@@ -15,7 +15,6 @@
 //= require lodash/lodash.min
 //= require uppy/dist/uppy.min
 //= require_tree .
-//= stub editor
 
 //FIXME: move to coffescript
 $(function(){
