+++ apps/dashboard/app/views/files/edit.html.erb
@@ -15,13 +15,6 @@
 
     <script src="<%= "#{ENV['RAILS_RELATIVE_URL_ROOT']}/ace/1.2.6/ace.js" %>" type="text/javascript" charset="utf-8"></script>
     <script src="<%= "#{ENV['RAILS_RELATIVE_URL_ROOT']}/ace/1.2.6/ext-modelist.js" %>" type="text/javascript" charset="utf-8"></script>
-    <script>
-      var editorContent = "";
-
-      const editorElement = document.querySelector('#editor');
-      const apiUrl = editorElement.dataset.api;
-      const filePath = editorElement.dataset.path
-    </script>
 
 <% elsif @directory_content %>
     <h3>Select a file to edit:</h3>
